Break Logo out
